requirements.txt
watchdog==2.1.6
pyyaml==6.1
requests==2.27.1
